refactor: modularize binary crate into focused modules
Split src/commands/root.rs into single-responsibility modules:
- src/cli.rs: CLI argument parsing, config conversion, and run()
- src/error.rs: application-level Error enum (renamed from RunError)
- src/server.rs: Handler dispatch, ServerHandler impl, create_handler()
- src/commands/: transport modules (http.rs, stdio.rs)
